+++ app/src/main/java/com/cajetan/youtubeplus/fragments/SearchFragment.kt
@@ -14,33 +14,32 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import com.cajetan.youtubeplus.R
 import com.cajetan.youtubeplus.adapters.ContentListAdapter
+import com.cajetan.youtubeplus.data.MainDataViewModel
 import com.cajetan.youtubeplus.data.PlaylistData
 import com.cajetan.youtubeplus.data.VideoData
-import com.cajetan.youtubeplus.data.MainDataViewModel
 import com.cajetan.youtubeplus.utils.FeedItem
 import com.cajetan.youtubeplus.utils.ItemType
 import com.cajetan.youtubeplus.utils.YouTubeData
-import com.cajetan.youtubeplus.viewmodels.StartViewModel
-import com.google.api.services.youtube.model.Video
 import org.jetbrains.anko.alert
+import org.jetbrains.anko.find
 import org.jetbrains.anko.noButton
 import org.jetbrains.anko.yesButton
 
-class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
-        YouTubeData.MostPopularListener {
-
-    private val TAG: String = this.javaClass.simpleName
+class SearchFragment : Fragment(), ContentListAdapter.ListItemClickListener,
+        YouTubeData.VideoSearchListener, YouTubeData.UploadPlaylistListener {
 
     private lateinit var mAdapter: ContentListAdapter
     private lateinit var mYouTubeData: YouTubeData
 
-    // TODO: view model in activity scope
     private lateinit var mMainDataViewModel: MainDataViewModel
-    private lateinit var mStartViewModel: StartViewModel
 
+    private lateinit var rootView: View
     private lateinit var videoList: RecyclerView
-    private lateinit var progressBarCentre: ProgressBar
-    private lateinit var progressBarBottom: ProgressBar
+    private lateinit var searchProgressBarCentre: ProgressBar
+    private lateinit var searchProgressBarBottom: ProgressBar
+
+    private var searchQuery: String = ""
+    private var nextPageToken: String = ""
 
     ////////////////////////////////////////////////////////////////////////////////
     // Lifecycle
@@ -50,17 +49,15 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
         super.onCreate(savedInstanceState)
         setHasOptionsMenu(true)
         mMainDataViewModel = ViewModelProviders.of(this).get(MainDataViewModel::class.java)
-
-        // TODO: experiment with parent activity as a context
-        mStartViewModel = ViewModelProviders.of(this).get(StartViewModel::class.java)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val view = inflater.inflate(R.layout.fragment_start, container, false)
 
+        rootView = view.findViewById(R.id.root_view)
         videoList = view.findViewById(R.id.videoList)
-        progressBarCentre = view.findViewById(R.id.progressBarCentre)
-        progressBarBottom = view.findViewById(R.id.progressBarBottom)
+        searchProgressBarCentre = view.findViewById(R.id.progressBarCentre)
+        searchProgressBarBottom = view.findViewById(R.id.progressBarBottom)
 
         return view
     }
@@ -68,14 +65,14 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
 
-        mAdapter = ContentListAdapter(mStartViewModel.getAdapterItems(), this, activity!!)
+        mAdapter = ContentListAdapter(emptyList(), this, activity!!)
         mYouTubeData = YouTubeData(activity!!, this)
 
-        setupVideoList()
-        // TODO: default argument should be used here
-        loadMostPopularVideos(mStartViewModel.nextPageToken)
+        setupSearchResultList()
+        searchVideos(arguments?.getString("search_query")!!)
     }
 
+    // TODO: fix search view focus
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.search_options_menu, menu)
 
@@ -89,8 +86,7 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     override fun onPrepareOptionsMenu(menu: Menu) {
         super.onPrepareOptionsMenu(menu)
         val searchView = menu.findItem(R.id.search)?.actionView as SearchView
-        searchView.setQuery("", false)
-        searchView.isIconified = true
+        searchView.isIconified = false
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean = when (item.itemId) {
@@ -100,7 +96,7 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
         }
 
         R.id.action_where_do_we_go -> {
-            findNavController().navigate(R.id.action_start_to_playerActivity,
+            findNavController().navigate(R.id.action_searchFragment_to_playerActivity,
                     bundleOf(getString(R.string.video_id_key) to getString(R.string.where_do_we_go_id)))
             true
         }
@@ -117,9 +113,9 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     // Init
     ////////////////////////////////////////////////////////////////////////////////
 
-    private fun setupVideoList() {
+    private fun setupSearchResultList() {
         mAdapter.onBottomReached = {
-            loadMostPopularVideos(mStartViewModel.nextPageToken)
+            loadSearchResults(nextPageToken)
         }
 
         videoList.setHasFixedSize(true)
@@ -130,50 +126,62 @@ class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     // Utils
     ////////////////////////////////////////////////////////////////////////////////
 
-    private fun loadMostPopularVideos(nextPageToken: String) {
+    fun searchVideos(query: String) {
+        if (query.isEmpty() || query == searchQuery)
+            return
+
+        searchQuery = query
+        loadSearchResults("")
+    }
+
+    private fun loadSearchResults(nextPageToken: String) {
         if (nextPageToken.isEmpty()) {
-            progressBarCentre.visibility = View.VISIBLE
+            searchProgressBarCentre.visibility = View.VISIBLE
             videoList.visibility = View.INVISIBLE
         } else {
-            progressBarBottom.visibility = View.VISIBLE
+            searchProgressBarBottom.visibility = View.VISIBLE
         }
 
-        mYouTubeData.receiveMostPopularResults(nextPageToken)
+        mYouTubeData.receiveSearchResults(searchQuery, nextPageToken)
     }
+
 
     ////////////////////////////////////////////////////////////////////////////////
     // Callbacks
     ////////////////////////////////////////////////////////////////////////////////
 
-    override fun onMostPopularReceived(results: List<Video>,
-                                       nextPageToken: String, previousPageToken: String) {
-        // TODO: possible reductions in logic here
+    override fun onUploadPlaylistIdReceived(id: String, channelTitle: String) {
+        findNavController().navigate(R.id.action_searchFragment_to_playlistContent,
+                bundleOf(getString(R.string.playlist_id_key) to id,
+                        getString(R.string.channel_title_key) to channelTitle))
+    }
+
+
+    override fun onSearchResultsReceived(results: List<FeedItem>,
+                                         nextPageToken: String, previousPageToken: String) {
         if (previousPageToken.isEmpty()) {
-            mStartViewModel.clearAdapterItems()
+            mAdapter.clearItems()
             videoList.scrollToPosition(0)
 
-            progressBarCentre.visibility = View.INVISIBLE
+            searchProgressBarCentre.visibility = View.INVISIBLE
             videoList.visibility = View.VISIBLE
         } else {
-            progressBarBottom.visibility = View.GONE
+            searchProgressBarBottom.visibility = View.GONE
         }
 
-        mStartViewModel.addAdapterItems(results.map { FeedItem(it.id, video = it) })
-        mStartViewModel.nextPageToken = nextPageToken
-
-        // TODO: observable LiveData
-        mAdapter.setItems(mStartViewModel.getAdapterItems())
+        mAdapter.addItems(results)
+        this.nextPageToken = nextPageToken
     }
 
     override fun onListItemClick(id: String, position: Int, type: ItemType) {
         when (type) {
-            ItemType.Video -> findNavController().navigate(R.id.action_start_to_playerActivity,
+            ItemType.Video -> findNavController().navigate(R.id.action_searchFragment_to_playerActivity,
                     bundleOf(getString(R.string.video_id_key) to id))
 
-            ItemType.Playlist -> findNavController().navigate(R.id.action_start_to_playerActivity,
+            ItemType.Playlist -> findNavController().navigate(R.id.action_searchFragment_to_playerActivity,
                     bundleOf(getString(R.string.playlist_id_key) to id))
 
-            else -> {}
+            ItemType.Channel -> mYouTubeData.receiveUploadPlaylistId(id)
         }
     }